Finish access tokens

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,8 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskStatus } from './task-status.enum';
 import { TasksModule } from './tasks.module';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from 'src/auth/user.entity';
+import { UserRepository } from 'src/auth/user.repository';
 // import { CreateTaskDto } from './dto/create-task.dto';
 // import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 
@@ -16,35 +18,38 @@ export class TasksService {
     private taskRepository: TaskRepository
   ) {}  
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user:User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    // select * from tasks where id = id and userId = user.id
+    console.log('user', user);
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!found) {
-      throw new NotFoundException(`Task with id ${id} not found`);
+      throw new NotFoundException(`Task with id ${id} not found for user ${user.username}`);
     }
 
     return found;
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-   const result = await this.taskRepository.delete(id);
-   if (result.affected === 0) {
-    throw new NotFoundException(`Task with id ${id} not found`);
-   }
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    // delete based on task id as well as user id 
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+    if (result.affected === 0) {
+      throw new NotFoundException(`Task with id ${id} not found for user ${user.username}`);
+    }
   }
 
-  async updateTaskById(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskById(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 }
